API on good progress / almost finished.
GET, POST and PUT methods issues related to other models fixed.
DELETE methods are yet to be fixed: they are currently working except in case of Foreign Keys (search for cascade deleting). But nonethelss, DELETE methods aren't necessary.

All tests so far gave the expected result. Maybe some additionnal test could be useful.

The controllers have all been changed and written manually. The changes were brought to include data from other model (to the profession of a user while getting information of a user for instance) and be able to PUT/POST without having to fill every model attributes of the database.

DTO were created based on models to select and specify the attributes from the models we wanted to keep and use in the API.

In the 'DataUsage' Model, the attribute 'UsedByNavigation' was called like this by the autocompleter. Since it refers to the users, it was change for User. Change was also done in the Context file.

The documentation remains to be written.